moved stronghold file, removed crypto folder, tested and build as well, all works
