--- identity-wallet/src/crypto/mod.rs
@@ -1,1 +1,0 @@
-pub mod stronghold;
