+++ identity-wallet/src/state/reducers/mod.rs
@@ -8,7 +8,7 @@ pub mod user_data_query;
 use super::actions::{listen, CancelUserFlow, SetLocale, UpdateCredentialMetadata, UpdateProfileSettings};
 use super::persistence::{clear_all_assets, delete_state_file, delete_stronghold, load_state};
 use super::{IdentityManager, ProfileSettings};
-use crate::crypto::stronghold::StrongholdManager;
+use crate::stronghold::StrongholdManager;
 use crate::error::AppError::{self, *};
 use crate::state::actions::{Action, CreateNew};
 use crate::state::user_prompt::CurrentUserPrompt;
