+++ identity-wallet/src/stronghold.rs
